+++ crates/lib/core/tests/crypto/mod.rs
@@ -1,9 +1,7 @@
-mod dsa;
 mod falcon;
 
 mod aead;
 mod blake3;
 mod circuit_evaluation;
-mod hashes;
 mod poseidon2;
 mod sha256;